naming and inheritance for importlib
This MR breaks out _frozen_importlib_external (which is the same
thing as importlib._bootstrap_external) and _frozen_importlib
(which is the same thing as importlib._bootstrap).

related to https://github.com/python/typeshed/issues/3968
